refactor: change dependency chain
