+++ packages/services/cdn-worker/src/artifact-storage-reader.ts
@@ -1,7 +1,4 @@
-/**
- * IMPORTANT NOTE: This file needs to be kept platform-agnostic, don't use any Node.js specific APIs.
- */
-import { AwsClient } from '../../../shared/aws';
+import { AwsClient } from './aws';
 
 const presignedUrlExpirationSeconds = 60;
 
@@ -17,32 +14,22 @@ export type ArtifactsType = SDLArtifactTypes | 'metadata' | 'services' | 'superg
  */
 export class ArtifactStorageReader {
   private publicUrl: URL | null;
-  private awsClient: AwsClient;
-  private s3Endpoint: string;
 
   constructor(
-    s3Config: {
-      accessKeyId: string;
-      secretAccessKey: string;
+    private s3: {
+      client: AwsClient;
       endpoint: string;
+      bucketName: string;
     },
-    private bucketName: string,
     /** The public URL in case the public S3 endpoint differs from the internal S3 endpoint. E.g. within a docker network. */
     publicUrl: string | null,
   ) {
     this.publicUrl = publicUrl ? new URL(publicUrl) : null;
-    this.awsClient = new AwsClient({
-      accessKeyId: s3Config.accessKeyId,
-      secretAccessKey: s3Config.secretAccessKey,
-      region: 'auto',
-      service: 's3',
-    });
-    this.s3Endpoint = s3Config.endpoint;
   }
 
   private async generatePresignedGetUrl(key: string): Promise<string> {
-    const signedUrl = await this.awsClient.sign(
-      this.s3Endpoint + `/` + this.bucketName + '/' + key,
+    const signedUrl = await this.s3.client.sign(
+      [this.s3.endpoint, this.s3.bucketName, key].join('/'),
       {
         method: 'GET',
         aws: { signQuery: true },
@@ -76,8 +63,8 @@ export class ArtifactStorageReader {
 
     const key = buildArtifactStorageKey(targetId, artifactType);
 
-    const response = await this.awsClient.fetch(
-      this.s3Endpoint + '/' + this.bucketName + '/' + key,
+    const response = await this.s3.client.fetch(
+      [this.s3.endpoint, this.s3.bucketName, key].join('/'),
       {
         method: 'HEAD',
       },
