+++ packages/services/cdn-worker/src/dev.ts
@@ -1,13 +1,30 @@
 import { createServer } from 'http';
 import itty from 'itty-router';
 import { json, withParams } from 'itty-router-extras';
-import { ArtifactStorageReader } from '@hive/api/src/modules/schema/providers/artifact-storage-reader';
 import { createServerAdapter } from '@whatwg-node/server';
 import { createArtifactRequestHandler } from './artifact-handler';
+import { ArtifactStorageReader } from './artifact-storage-reader';
+import { AwsClient } from './aws';
 import './dev-polyfill';
 import { devStorage } from './dev-polyfill';
 import { createRequestHandler } from './handler';
 import { createIsKeyValid } from './key-validation';
+
+declare let S3_ENDPOINT: string;
+declare let S3_ACCESS_KEY_ID: string;
+declare let S3_SECRET_ACCESS_KEY: string;
+declare let S3_BUCKET_NAME: string;
+declare let S3_PUBLIC_URL: string;
+
+const s3 = {
+  client: new AwsClient({
+    accessKeyId: S3_ACCESS_KEY_ID,
+    secretAccessKey: S3_SECRET_ACCESS_KEY,
+    service: 's3',
+  }),
+  bucketName: S3_BUCKET_NAME,
+  endpoint: S3_ENDPOINT,
+};
 
 // eslint-disable-next-line no-process-env
 const PORT = process.env.PORT ? parseInt(process.env.PORT, 10) : 4010;
@@ -29,21 +46,7 @@ const handleRequest = createRequestHandler({
   isKeyValid: createIsKeyValid({ keyData: KEY_DATA }),
 });
 
-declare let S3_ENDPOINT: string;
-declare let S3_ACCESS_KEY_ID: string;
-declare let S3_SECRET_ACCESS_KEY: string;
-declare let S3_BUCKET_NAME: string;
-declare let S3_PUBLIC_URL: string;
-
-const artifactStorageReader = new ArtifactStorageReader(
-  {
-    accessKeyId: S3_ACCESS_KEY_ID,
-    secretAccessKey: S3_SECRET_ACCESS_KEY,
-    endpoint: S3_ENDPOINT,
-  },
-  S3_BUCKET_NAME,
-  S3_PUBLIC_URL,
-);
+const artifactStorageReader = new ArtifactStorageReader(s3, S3_PUBLIC_URL);
 
 const handleArtifactRequest = createArtifactRequestHandler({
   isKeyValid: createIsKeyValid({ keyData: KEY_DATA }),
