+++ packages/services/cdn-worker/src/index.ts
@@ -1,11 +1,29 @@
 import itty from 'itty-router';
 import Toucan from 'toucan-js';
-import { ArtifactStorageReader } from '@hive/api/src/modules/schema/providers/artifact-storage-reader';
 import { AnalyticsEngine, createAnalytics } from './analytics';
 import { createArtifactRequestHandler } from './artifact-handler';
+import { ArtifactStorageReader } from './artifact-storage-reader';
+import { AwsClient } from './aws';
 import { UnexpectedError } from './errors';
 import { createRequestHandler } from './handler';
 import { createIsKeyValid } from './key-validation';
+
+declare let S3_ENDPOINT: string;
+declare let S3_ACCESS_KEY_ID: string;
+declare let S3_SECRET_ACCESS_KEY: string;
+declare let S3_BUCKET_NAME: string;
+
+const s3 = {
+  client: new AwsClient({
+    accessKeyId: S3_ACCESS_KEY_ID,
+    secretAccessKey: S3_SECRET_ACCESS_KEY,
+    service: 's3',
+  }),
+  bucketName: S3_BUCKET_NAME,
+  endpoint: S3_ENDPOINT,
+};
+
+const artifactStorageReader = new ArtifactStorageReader(s3, null);
 
 /**
  * KV Storage for the CDN
@@ -30,76 +48,61 @@ declare let SENTRY_RELEASE: string;
 declare let USAGE_ANALYTICS: AnalyticsEngine;
 declare let ERROR_ANALYTICS: AnalyticsEngine;
 
-const isKeyValid = createIsKeyValid({ keyData: KEY_DATA });
-
-declare let S3_ENDPOINT: string;
-declare let S3_ACCESS_KEY_ID: string;
-declare let S3_SECRET_ACCESS_KEY: string;
-declare let S3_BUCKET_NAME: string;
-
 const analytics = createAnalytics({
   usage: USAGE_ANALYTICS,
   error: ERROR_ANALYTICS,
 });
 
-const handleRequest = createRequestHandler({
-  getRawStoreValue: value => HIVE_DATA.get(value),
-  isKeyValid,
-  analytics,
-});
-
-const artifactStorageReader = new ArtifactStorageReader(
-  {
-    accessKeyId: S3_ACCESS_KEY_ID,
-    secretAccessKey: S3_SECRET_ACCESS_KEY,
-    endpoint: S3_ENDPOINT,
-  },
-  S3_BUCKET_NAME,
-  null,
-);
-
-const handleArtifactRequest = createArtifactRequestHandler({
-  isKeyValid,
-  analytics,
-  async getArtifactAction(targetId, artifactType, eTag) {
-    return artifactStorageReader.generateArtifactReadUrl(targetId, artifactType, eTag);
-  },
-  async fallback(request: Request, params: { targetId: string; artifactType: string }) {
-    const artifactTypeMap: Record<string, string> = {
-      metadata: 'metadata',
-      sdl: 'sdl',
-      services: 'schema',
-      supergraph: 'supergraph',
-    };
-    const artifactType = artifactTypeMap[params.artifactType];
-
-    if (artifactType) {
-      const url = request.url.replace(
-        `/artifacts/v1/${params.targetId}/${params.artifactType}`,
-        `/${params.targetId}/${artifactType}`,
-      );
-
-      return handleRequest(new Request(url, request));
-    }
-
-    return;
-  },
-});
-
-const router = itty
-  .Router()
-  .get(
-    '/_health',
-    () =>
-      new Response('OK', {
-        status: 200,
-      }),
-  )
-  .get('*', handleArtifactRequest)
-  // Legacy CDN Handlers
-  .get('*', handleRequest);
-
 self.addEventListener('fetch', async (event: FetchEvent) => {
+  const isKeyValid = createIsKeyValid({ keyData: KEY_DATA });
+
+  const handleRequest = createRequestHandler({
+    getRawStoreValue: value => HIVE_DATA.get(value),
+    isKeyValid,
+    analytics,
+  });
+
+  const handleArtifactRequest = createArtifactRequestHandler({
+    isKeyValid,
+    analytics,
+    async getArtifactAction(targetId, artifactType, eTag) {
+      return artifactStorageReader.generateArtifactReadUrl(targetId, artifactType, eTag);
+    },
+    async fallback(request: Request, params: { targetId: string; artifactType: string }) {
+      const artifactTypeMap: Record<string, string> = {
+        metadata: 'metadata',
+        sdl: 'sdl',
+        services: 'schema',
+        supergraph: 'supergraph',
+      };
+      const artifactType = artifactTypeMap[params.artifactType];
+
+      if (artifactType) {
+        const url = request.url.replace(
+          `/artifacts/v1/${params.targetId}/${params.artifactType}`,
+          `/${params.targetId}/${artifactType}`,
+        );
+
+        return handleRequest(new Request(url, request));
+      }
+
+      return;
+    },
+  });
+
+  const router = itty
+    .Router()
+    .get(
+      '/_health',
+      () =>
+        new Response('OK', {
+          status: 200,
+        }),
+    )
+    .get('*', handleArtifactRequest)
+    // Legacy CDN Handlers
+    .get('*', handleRequest);
+
   const sentry = new Toucan({
     dsn: SENTRY_DSN,
     environment: SENTRY_ENVIRONMENT,
