+++ tsconfig.json
@@ -33,6 +33,10 @@
       "@hive/cdn-script/artifact-handler": [
         "./packages/services/cdn-worker/src/artifact-handler.ts"
       ],
+      "@hive/cdn-script/artifact-storage-reader": [
+        "./packages/services/cdn-worker/src/artifact-storage-reader.ts"
+      ],
+      "@hive/cdn-script/aws": ["./packages/services/cdn-worker/src/aws.ts"],
       "@hive/cdn-script/key-validation": ["./packages/services/cdn-worker/src/key-validation.ts"],
       "@hive/server": ["./packages/services/server/src/api.ts"],
       "@hive/stripe-billing": ["./packages/services/stripe-billing/src/api.ts"],
